+++ src/components/ActivityCard/styles.ts
