+++ src/components/ActivityCardMap/styles.ts
